changes to add plans and subscriptions

+++ app/Http/Controllers/PlanController.php
@@ -4,12 +4,21 @@ namespace App\Http\Controllers;
 
 use App\Http\Resources\Plans\PlanDetailResource;
 use App\Http\Resources\Plans\PlanListResource;
+use App\Models\Plan;
 use App\Models\User;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Auth;
 
 class PlanController extends Controller
 {
+    private function authorizeRoot()
+    {
+        if (optional(Auth::user()->contact)->firstRole() !== 'ROOT') {
+            return response()->json(['message' => 'User does not have permission to access plans.'], 422);
+        }
+        return null;
+    }
+
     /**
      * Display a listing of the resource.
      *
@@ -17,6 +26,9 @@ class PlanController extends Controller
      */
     public function index()
     {
+        if ($response = $this->authorizeRoot()) {
+            return $response;
+        }
         $plans = Plan::all();
         return PlanListResource::collection($plans);
     }
@@ -30,6 +42,10 @@ class PlanController extends Controller
      */
     public function store(Request $request)
     {
+        if ($response = $this->authorizeRoot()) {
+            return $response;
+        }
+
         // Validate the request
         $request->validate([
            'name' => 'string',
@@ -52,6 +68,10 @@ class PlanController extends Controller
      */
     public function show($plan)
     {
+        if ($response = $this->authorizeRoot()) {
+            return $response;
+        }
+
         return response()->json($plan);
         return  new PlanDetailResource($plan);
     }
@@ -66,6 +86,10 @@ class PlanController extends Controller
      */
     public function update(Request $request, $plan)
     {
+        if ($response = $this->authorizeRoot()) {
+            return $response;
+        }
+
          // Validate the request
          $request->validate([
             'name' => 'string',
@@ -88,6 +112,10 @@ class PlanController extends Controller
      */
     public function destroy($plan)
     {
+        if ($response = $this->authorizeRoot()) {
+            return $response;
+        }
+
         $plan->delete();
         return response()->json(['message' => 'Plan deleted successfully'], 204);
     }
